move SUEPNet into inference_utils
